refactor: LottoResult 로또당첨 결과 도메인 의존성 제거
